Money, multiple item gets/drops, tab fixes, other bug fixes.

+++ src/KatanaMUD/Models/Currency.cs
@@ -14,7 +14,7 @@ namespace KatanaMUD.Models
         private String _Name;
         private String _ShortName;
         private Int64 _Value;
-        private float _Weight;
+        private double _Weight;
 
         public Currency()
         {
@@ -24,7 +24,7 @@ namespace KatanaMUD.Models
         public String Name { get { return _Name; } set { _Name = value; this.Changed(); } }
         public String ShortName { get { return _ShortName; } set { _ShortName = value; this.Changed(); } }
         public Int64 Value { get { return _Value; } set { _Value = value; this.Changed(); } }
-        public float Weight { get { return _Weight; } set { _Weight = value; this.Changed(); } }
+        public double Weight { get { return _Weight; } set { _Weight = value; this.Changed(); } }
         public static Currency Load(SqlDataReader reader)
         {
             var entity = new Currency();
@@ -32,7 +32,7 @@ namespace KatanaMUD.Models
             entity._Name = reader.GetString(1);
             entity._ShortName = reader.GetString(2);
             entity._Value = reader.GetInt64(3);
-            entity._Weight = reader.GetFloat(4);
+            entity._Weight = reader.GetDouble(4);
             return entity;
         }
